fix eval primitives

+++ rispy/src/expr.rs
@@ -1,7 +1,7 @@
 #[derive(Debug, PartialEq, Clone)]
 pub enum Expr {
     List(Vec<Expr>),
-    Num(i32),
+    Num(f64),
     Keyword(String),
     Boolean(bool),
     Quote(Vec<Expr>),
